requirements.txt
# core
streamlit>=1.20
fastapi>=0.95
uvicorn[standard]>=0.22
pandas
pydantic
tf-keras

# vector + embeddings
faiss-cpu>=1.7.4   # or faiss-gpu if you have CUDA
sentence-transformers>=2.2.2
numpy

# LLM clients (pick whichever you use)
google-generativeai>=0.7.0  # if using Gemini
openai>=1.0.0                # optional fallback

# helpers
python-dotenv
rich
tqdm
dotenv
